microservices/auth-service/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
prometheus-fastapi-instrumentator
python-json-logger
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
alembic
redis
google-auth
requests
